Add operator functions to built-in types
This may not be the best approach but it works and is generic enough to
convey all the types that has some of the operations defined. Apparently
there's no way to know which types accept which operators.

In the future, might be worth it to isolate each operation in its own
bound function.